+++ ktlint-test/src/main/kotlin/com/pinterest/ktlint/test/RuleExtension.kt
@@ -5,6 +5,7 @@ import com.pinterest.ktlint.core.LintError
 import com.pinterest.ktlint.core.Rule
 import com.pinterest.ktlint.core.RuleSet
 import java.util.ArrayList
+import org.assertj.core.api.Assertions
 import org.assertj.core.util.diff.DiffUtils.diff
 import org.assertj.core.util.diff.DiffUtils.generateUnifiedDiff
 
@@ -105,7 +106,7 @@ public fun Rule.diffFileLint(
             diff(expected.map(str), actual.map(str)),
             expected.size + actual.size
         ).joinToString("\n")
-    return if (diff.isEmpty()) "" else diff
+    return diff.ifEmpty { "" }
 }
 
 public fun Rule.diffFileFormat(
@@ -118,7 +119,31 @@ public fun Rule.diffFileFormat(
     val diff =
         generateUnifiedDiff(expectedPath, "output", expected, diff(expected, actual), expected.size + actual.size)
             .joinToString("\n")
-    return if (diff.isEmpty()) "" else diff
+    return diff.ifEmpty { "" }
+}
+
+/**
+ * Alternative to [diffFileFormat]. Depending on your personal favor it might be more insightful whenever a test is
+ * failing. Currently it is offered as utility so it can be used during development.
+ *
+ * To be used as:
+ *
+ *     @Test
+ *     fun testFormatRawStringTrimIndent() {
+ *         IndentationRule().assertThatFileFormat(
+ *             "spec/indent/format-raw-string-trim-indent.kt.spec",
+ *             "spec/indent/format-raw-string-trim-indent-expected.kt.spec"
+ *         )
+ *     }
+ */
+public fun Rule.assertThatFileFormat(
+    srcPath: String,
+    expectedPath: String,
+    userData: Map<String, String> = emptyMap()
+) {
+    val actual = format(getResourceAsText(srcPath), userData, script = true).split('\n')
+    val expected = getResourceAsText(expectedPath).split('\n')
+    Assertions.assertThat(actual).isEqualTo(expected)
 }
 
 private fun getResourceAsText(path: String) =