+++ ktlint/src/main/kotlin/com/pinterest/ktlint/internal/RuleSetsLoader.kt
@@ -26,11 +26,33 @@ internal fun JarFiles.loadRulesets(
     .filterKeys { loadExperimental || it != "experimental" }
     .filterKeys { !(disabledRules.isStandardRuleSetDisabled() && it == "\u0000standard") }
     .toSortedMap()
-    .also {
+    .also { ruleSetMap ->
         if (debug) {
-            it.forEach { entry ->
+            ruleSetMap.forEach { entry ->
                 println("[DEBUG] Discovered ruleset with \"${entry.key}\" id.")
             }
+        }
+
+        val expectedNumberOfCustomRuleSetsToBeLoaded = this.distinct().count()
+        val customRuleSetsLoaded =
+            ruleSetMap
+                .filterKeys { it != "experimental" && it != "\u0000standard" }
+                .values
+                .map { it.javaClass.canonicalName }
+        val actualNumberOfCustomRuleSetsLoaded = customRuleSetsLoaded.count()
+        if (expectedNumberOfCustomRuleSetsToBeLoaded != actualNumberOfCustomRuleSetsLoaded) {
+            System.err.println(
+                """
+                [WARNING] Number of custom rule sets loaded does not match the expected number of rule sets to be loaded.
+                          Expected to load $expectedNumberOfCustomRuleSetsToBeLoaded custom rule set(s) for rules: ${this.joinToString()}
+                          Actually loaded $actualNumberOfCustomRuleSetsLoaded custom rule set(s) with names: ${customRuleSetsLoaded.joinToString() }
+                          One or more of the specified jars does not provide the custom rule set correctly. Check following:
+                            - Does the jar contain an implementation of the RuleSetProvider interface?
+                            - Does the jar contain a resource file with name "com.pinterest.ktlint.core.RuleSetProvider"?
+                            - Is the resource file located in directory "src/main/resources/META-INF/services"?
+                            - Does the resource file contain the fully qualified class name of the class implementing the RuleSetProvider interface?
+                """.trimIndent() // ktlint-disable string-template
+            )
         }
     }